+++ core/api/typescript/src/lib/enclaves/grpc_web_api_container_client.ts
@@ -22,7 +22,7 @@ import {
     ExecuteModuleArgs,
     ExecuteModuleResponse,
     ExecCommandArgs,
-    ExecCommandResponse,
+    ExecCommandResponse, StoreWebFilesArtifactResponse, StoreWebFilesArtifactArgs,
 } from "../../kurtosis_core_rpc_api_bindings/api_container_service_pb";
 import { ApiContainerServiceClient as ApiContainerServiceClientWeb } from "../../kurtosis_core_rpc_api_bindings/api_container_service_grpc_web_pb";
 import { GenericApiContainerClient } from "./generic_api_container_client";
@@ -362,5 +362,27 @@ export class GrpcWebApiContainerClient implements GenericApiContainerClient {
 
         const execCommandResponse = execCommandResponseResult.value
         return ok(execCommandResponse)
+    }
+
+    public async storeWebFilesArtifact(storeWebFilesArtifactArgs: StoreWebFilesArtifactArgs): Promise<Result<StoreWebFilesArtifactResponse, Error>> {
+        const storeWebFilesArtifactPromise: Promise<Result<StoreWebFilesArtifactResponse, Error>> = new Promise((resolve, _unusedReject) => {
+            this.client.storeWebFilesArtifact(storeWebFilesArtifactArgs, {}, (error: grpc_web.RpcError | null, response?: StoreWebFilesArtifactResponse) => {
+                if (error === null) {
+                    if (!response) {
+                        resolve(err(new Error("No error was encountered but the response was still falsy; this should never happen")));
+                    } else {
+                        resolve(ok(response!));
+                    }
+                } else {
+                    resolve(err(error));
+                }
+            })
+        });
+        const storeWebFilesArtifactResponseResult = await storeWebFilesArtifactPromise;
+        if (storeWebFilesArtifactResponseResult.isErr()) {
+            return err(storeWebFilesArtifactResponseResult.error)
+        }
+        const storeWebFilesArtifactResponse = storeWebFilesArtifactResponseResult.value;
+        return ok(storeWebFilesArtifactResponse);
     }
 }