+++ core/docs/changelog.md
@@ -7,6 +7,7 @@
 # 1.43.3
 ### Features
 * Add `EnclaveContext.WithFiles` for specifying mounting files artifacts that come back from the `EnclaveContext.UploadFiles` command
+* Add `EnclaveContext.StoreWebFiles` for downloading files artifacts from the internet
 
 # 1.43.2
 ### Fixes